Got everything in working order hopefully
