+++ lib/stat_tracker.rb
@@ -27,6 +27,11 @@ class StatTracker
     highest_total.away_goals + highest_total.home_goals
   end
 
+  def lowest_total_score
+    lowest_total = @games.games.min_by{ |game| game.away_goals + game.home_goals }
+    lowest_total.away_goals + lowest_total.home_goals
+  end
+
   def biggest_blowout
     blowout = @games.games.max_by { |game| (game.home_goals - game.away_goals).abs }
     (blowout.home_goals - blowout.away_goals).abs
@@ -34,17 +39,48 @@ class StatTracker
 
   def percent_home_wins
     h_win = @games.games.count { |game| game.away_goals < game.home_goals }
-    (h_win * 100.00 / @games.games.length).to_f
+    ((h_win * 100.00).to_f / @games.games.length).round(2)
   end
 
   def percent_visitor_wins
     v_win = @games.games.count { |game| game.away_goals > game.home_goals }
-    (v_win * 100.00 / @games.games.length).to_f
+    ((v_win * 100.00).to_f / @games.games.length).round(2)
   end
 
   def percent_ties
     ties = @games.games.count { |game| game.away_goals == game.home_goals }
-    (ties * 100.00 / @games.games.length).to_f
+    ((ties * 100.00).to_f / @games.games.length).round(2)
   end
+
+  def game_count_per_season
+    hash = @games.games.reduce({}) do |game_count, game|
+      if game_count[game.season]
+        game_count[game.season] += 1
+      else
+        game_count[game.season] = 1
+      end
+      game_count
+    end
+    hash
+  end
+
+def goal_count_per_season
+    hash = @games.games.reduce({}) do |goal_count, game|
+      if goal_count[game.season]
+        goal_count[game.season] += (game.away_goals + game.home_goals)
+      else
+        goal_count[game.season] = (game.away_goals + game.home_goals)
+      end
+      goal_count
+    end
+    hash
+  end
+
+  def average_goals_by_season
+    goal_count_per_season.merge(game_count_per_season) do |key, goal_count, game_count|
+      (goal_count.to_f / game_count).round(2)
+    end
+  end
+end
