--- stash/play.rb
@@ -1,19 +1,0 @@
-require 'pry'
-require 'CSV'
-
-games_array = []
-CSV.foreach("./data/games.csv", headers: true, header_converters: :symbol) do |row|
-  game_hash = {}
-  game_hash[:game_id] = row[:game_id].to_i
-  game_hash[:season] = row[:season].to_i
-  game_hash[:type] = row[:type]
-  game_hash[:date_time] = row[:date_time]
-  game_hash[:away_team_id] = row[:away_team_id].to_i
-  game_hash[:home_team_id] = row[:home_team_id].to_i
-  game_hash[:away_goals] = row[:away_goals].to_i
-  game_hash[:home_goals] = row[:home_goals].to_i
-  game_hash[:venue] = row[:venue]
-  game_hash[:venue_link] = row[:venue_link]
-games_array << game_hash
-end
-# p games_array
