+++ test/test_game_data.csv
@@ -4,7 +4,7 @@ game_id,season,type,date_time,away_team_id,home_team_id,away_goals,home_goals,ve
 2012030233,20122013,Postseason,5/20/13,16,17,1,3,Dignity Health Sports Park,/api/v1/venues/null
 2012030121,20122013,Postseason,5/2/13,9,8,2,2,Red Bull Arena,/api/v1/venues/null
 2012030323,20122013,Postseason,6/5/13,16,26,1,3,Nippert Stadium,/api/v1/venues/null
-2013030155,20132014,Postseason,4git br/27/14,30,21,3,2,BC Place,/api/v1/venues/null
+2013030155,20132014,Postseason,4/27/14,30,21,3,2,BC Place,/api/v1/venues/null
 2015030154,20152016,Postseason,4/21/16,25,30,3,2,Exploria Stadium,/api/v1/venues/null
 2016030135,20162017,Postseason,4/21/17,10,15,1,2,Providence Park,/api/v1/venues/null
 2016030162,20162017,Postseason,4/15/17,19,30,2,1,Exploria Stadium,/api/v1/venues/null
